get_languages.js adds languages to languages.json
The get_languages.js script can now list the supported language codes,
list languages in a given language and create or add languages to the
languages.json file. Because of the translations the format of the file
is slightly modified, see line 149 in the script. The rtl attributes
must be added manually.