Fix missing language translations bug
The bug was that if Firefox's locale (browser.useragent.locale in
about:config) was such that there weren't translations for that locale
in data/languages.json file, gtranslate's menucontext didn't show up at
all.

To fix this, a support for a locale is tested and in case a locale isn't
supported, a fallback locale is used.

See PR #77 for additional conversation.